Add Markdown support (Issue #302)

+++ htmldoc/zipc.c
@@ -889,7 +889,7 @@ zipc_write_local_trailer(
     * Update the CRC-32, compressed size, and uncompressed size fields...
     */
 
-    fseek(zc->fp, zf->offset + 14, SEEK_SET);
+    fseek(zc->fp, (long)(zf->offset + 14), SEEK_SET);
 
     status |= zipc_write_u32(zc, zf->crc32);
     status |= zipc_write_u32(zc, (unsigned)zf->compressed_size);
